Use clone to install lib dependency in sketch compilation workflow
The "Compile Examples" GitHub Actions workflow defines the dependencies of the example sketches which must be defined before attempting their compilation.

Previously, it attempted to install the "ESP Async WebServer" via Library Manager, but the library maintainer never submitted the library to Library Manager so that is not possible. Fortunately, the `arduino/compile-sketches` GitHub Actions action supports installing dependencies from multiple sources, which include Git repositories.

This new configuration will cause the library to be installed from a clone of the `https://github.com/me-no-dev/ESPAsyncWebServer.git` repository, checked out to the tip of the default branch.